+++ src/supadl/transfer/__init__.py
@@ -1,0 +1,11 @@
+"""HTTP transfer infrastructure and metadata probing."""
+
+from supadl.transfer.client import DEFAULT_USER_AGENT, create_http_client
+from supadl.transfer.probe import ProbeResult, ProbeService
+
+__all__ = [
+    "DEFAULT_USER_AGENT",
+    "ProbeResult",
+    "ProbeService",
+    "create_http_client",
+]